Changes wrong variable types
Revert "Fixes using Py_ssize_t rather than int"

This reverts commit 1e50d68250c0a52a2a269ba1a5cb27fb14d620ce.

+++ Src/_bindexmodule.c
@@ -258,10 +258,10 @@ _bindex_bindex (PyObject* unused, PyObject* args) {
     int eLONS[1];
     PyObject* HEAD;
     PyArrayObject* aHEAD;
-    Py_ssize_t eHEAD[1];
+    int eHEAD[1];
     PyObject* NEXT;
     PyArrayObject* aNEXT;
-    Py_ssize_t eNEXT[1];
+    int eNEXT[1];
     aLATS = (PyArrayObject*) 0;
     aLONS = (PyArrayObject*) 0;
     aHEAD = (PyArrayObject*) 0;
